대여 반납 기능 구현 완료 유효성 검사 필요

+++ sakila/src/main/java/com/gd/sakila/mapper/SalesMapper.java
@@ -10,7 +10,7 @@ public interface SalesMapper {
 	List<Map<String,Object>> selectBestSellerTop10();
 	List<Map<String,Object>> selectSalesByCategoryList();
 	List<Map<String,Object>> selectmonthlySalesList();
-	int selectTodaySales();
+	double selectTodaySales();
 	int selectTodayRental();
 	int selectRentalsCount();
 	int selectOverdueCount();
